fix(mcp-apps): self-review — one AB4772 per substituted view, honest fallback over the host bound, shared Rspack stats parser on agent-bundle/api, orphan-aware largest-modules ranking; rsc-agent-runtime silent dev log level (#585)
