+++ packages/agent-bundle/tests/mcp-apps-compile.test.ts
@@ -3,24 +3,11 @@ import { tmpdir } from 'node:os';
 import { dirname, join } from 'node:path';
 
 import { afterEach, describe, expect, it } from '@rstest/core';
-import type { Rspack } from '@rsbuild/core';
 
-import {
-  largestModules,
-  MCP_APP_COMPILE_ERROR_CAP,
-  MCP_APP_HTML_ADVISORY_BYTES,
-  mcpAppCompileErrorDiagnostics,
-  mcpAppCompileWarningDiagnostics,
-  mcpAppSizeDiagnostic,
-  normalizeStatsMessage,
-  statsErrorFile,
-  statsErrorLocation,
-  type McpAppDiagnosticContext,
-} from '../src/build/mcp-app-diagnostics.ts';
+import { MCP_APP_HTML_ADVISORY_BYTES } from '../src/build/mcp-app-diagnostics.ts';
 import { compileMcpApps, composeMcpAppsRsbuildConfig, type McpAppCompileMode } from '../src/build/mcp-apps.ts';
 import { DiagnosticError, type Diagnostic } from '../src/core/diagnostics.ts';
 import { MAX_APP_HTML_BYTES } from '../src/core/mcp-app-limits.ts';
-import { formatByteSize } from '../src/core/strings.ts';
 import type { AgentBundleToolsConfig, NormalizedMcpApp } from '../src/core/types.ts';
 import type { AgentBundleMeta } from '../src/meta.ts';
 import { workbenchNodeModules } from './helpers/workspace-paths.ts';
@@ -288,15 +275,26 @@ describe('compileMcpApps', () => {
     })).rejects.toThrow(/beyond the stable self-contained MCP App HTML output: static\/js\/status\.js\.map\. Only inline source maps/u);
   }, 60_000);
 
+  /**
+   * Exported functions nobody imports: readable output keeps every one (long
+   * names, comments and all — about 3 MiB for 16,000), the production
+   * minifier drops them all.
+   */
+  const readableOnlyPadding = Array.from({ length: 16_000 }, (_, index) => [
+    `/** Padding function number ${String(index)} that the readable build keeps and the minifier removes. */`,
+    `export function paddingFunctionWithAVeryLongName${String(index)}(): string { return 'padding-${String(index)}'; }`,
+  ].join('\n')).join('\n');
+
+  /**
+   * A string literal of `bytes` that no profile can shrink: the same size in
+   * readable and production output — as long as the view uses the string
+   * itself; the minifier folds `blob.length` to a number and drops it.
+   */
+  const incompressibleModule = (bytes: number): string => `export const blob = '${'x'.repeat(bytes)}';\n`;
+
   it('falls back to the production profile in development when the readable output would not render in the hosts', async () => {
-    // Exported functions nobody imports: readable output keeps every one
-    // (long names, comments and all), the production minifier drops them.
-    const filler = Array.from({ length: 16_000 }, (_, index) => [
-      `/** Padding function number ${String(index)} that the readable build keeps and the minifier removes. */`,
-      `export function paddingFunctionWithAVeryLongName${String(index)}(): string { return 'padding-${String(index)}'; }`,
-    ].join('\n')).join('\n');
     const files = {
-      'views/padding.ts': `${filler}\nexport const marker = 'fallback-ready';\n`,
+      'views/padding.ts': `${readableOnlyPadding}\nexport const marker = 'fallback-ready';\n`,
       'views/status.ts': "import { marker } from './padding';\ndocument.body.textContent = marker;\n",
     };
     const root = await createProject(files);
@@ -314,10 +312,76 @@ describe('compileMcpApps', () => {
       severity: 'warning',
       sourcePath: join(root, 'views', 'status.ts'),
     })]);
+    // The substitution notice names where the readable bytes came from.
     expect(result.diagnostics[0]!.message).toMatch(
-      /^MCP App "status" readable development output compiled to \d+(?:\.\d)? MiB, above the 2 MiB bound the Workbench and serve-app hosts accept; the preview renders the production build \(\d+ B, \d+ B gzip\) instead$/u,
+      /^MCP App "status" readable development output compiled to \d+(?:\.\d)? MiB, above the 2 MiB bound the Workbench and serve-app hosts accept; the preview renders the production build \(\d+ B, \d+ B gzip\) instead; largest modules: views\/padding\.ts \(\d+(?:\.\d)? MiB\), /u,
+    );
+    expect(result.diagnostics[0]!.message).toContain('views/status.ts (');
+  }, 120_000);
+
+  it('reports one AB4772 for a substituted view whose production build itself draws the advisory', async () => {
+    // Readable output: padding plus the blob, past 2 MiB. Production: the
+    // blob alone, between the 1 MiB advisory and the 2 MiB host bound.
+    const files = {
+      'views/blob.ts': incompressibleModule(1_200_000),
+      'views/padding.ts': `${readableOnlyPadding}\nexport const marker = 'fallback-ready';\n`,
+      'views/status.ts': "import { blob } from './blob';\nimport { marker } from './padding';\ndocument.title = marker;\ndocument.body.textContent = blob;\n",
+    };
+    const root = await createProject(files);
+    const { outDir, result } = await compile(root, [app(root)], { mode: 'development' });
+    const bytes = Buffer.byteLength(await emittedHtml(outDir), 'utf8');
+    expect(bytes).toBeGreaterThanOrEqual(MCP_APP_HTML_ADVISORY_BYTES);
+    expect(bytes).toBeLessThanOrEqual(MAX_APP_HTML_BYTES);
+    expect(result.apps[0]!.size.bytes).toBe(bytes);
+    // Exactly one size diagnostic: the substitution notice, not also the
+    // production compile's own 1 MiB advisory for the same view.
+    expect(result.diagnostics.map((diagnostic) => diagnostic.code)).toEqual(['AB4772']);
+    expect(result.diagnostics[0]!.message).toMatch(
+      /^MCP App "status" readable development output compiled to \d+(?:\.\d)? MiB, above the 2 MiB bound the Workbench and serve-app hosts accept; the preview renders the production build \(1\.1 MiB, \d+(?:\.\d)? KiB gzip\) instead; largest modules: views\/padding\.ts \(\d+(?:\.\d)? MiB\), views\/blob\.ts \(1\.1 MiB\)/u,
     );
   }, 120_000);
+
+  it('does not claim a substitution when the production build would not render in the hosts either', async () => {
+    const files = {
+      'views/blob.ts': incompressibleModule(2_200_000),
+      'views/status.ts': "import { blob } from './blob';\ndocument.body.textContent = blob;\n",
+    };
+    const root = await createProject(files);
+    const { outDir, result } = await compile(root, [app(root)], { mode: 'development' });
+    const html = await emittedHtml(outDir);
+    // The smaller production document is what lands, and it still does not fit.
+    expect(html).not.toContain('CONCATENATED MODULE');
+    expect(Buffer.byteLength(html, 'utf8')).toBeGreaterThan(MAX_APP_HTML_BYTES);
+    expect(result.apps[0]!.size.bytes).toBe(Buffer.byteLength(html, 'utf8'));
+    expect(result.diagnostics).toEqual([expect.objectContaining({
+      code: 'AB4772',
+      recovery: 'Trim the largest modules listed and rebuild; the Workbench and serve-app hosts refuse a view above 2 MiB.',
+      severity: 'warning',
+      sourcePath: join(root, 'views', 'status.ts'),
+    })]);
+    // The plain host-bound advisory for the production bytes — no "the preview renders" claim.
+    expect(result.diagnostics[0]!.message).toMatch(
+      /^MCP App "status" compiled to 2\.1 MiB \(\d+(?:\.\d)? KiB gzip\), above the 2 MiB bound the Workbench and serve-app hosts accept — the view will not render there; largest modules: views\/blob\.ts \(2\.1 MiB\), /u,
+    );
+    expect(result.diagnostics[0]!.message).toContain('views/status.ts (');
+    expect(result.diagnostics[0]!.message).not.toContain('preview renders');
+  }, 120_000);
+
+  it('names the largest authored modules in a production advisory, concatenated or not', async () => {
+    // Concatenation folds the author's ESM modules into the entry; the
+    // advisory still has to name the part that carries the bytes.
+    const files = {
+      'views/blob.ts': incompressibleModule(1_200_000),
+      'views/status.ts': "import { blob } from './blob';\ndocument.body.textContent = blob;\n",
+    };
+    const root = await createProject(files);
+    const { result } = await compile(root, [app(root)]);
+    expect(result.apps[0]!.size.bytes).toBeGreaterThanOrEqual(MCP_APP_HTML_ADVISORY_BYTES);
+    expect(result.diagnostics).toEqual([expect.objectContaining({ code: 'AB4772', severity: 'warning' })]);
+    expect(result.diagnostics[0]!.message).toMatch(
+      /^MCP App "status" compiled to 1\.1 MiB \(\d+(?:\.\d)? KiB gzip\), above the 1 MiB advisory bound; largest modules: views\/blob\.ts \(1\.1 MiB\), views\/status\.ts \(\d+ B\)/u,
+    );
+  }, 60_000);
 });
 
 describe('composeMcpAppsRsbuildConfig', () => {
@@ -339,206 +403,5 @@ describe('composeMcpAppsRsbuildConfig', () => {
     expect(development.output?.sourceMap).toBe(false);
     expect(development.output?.minify).toBe(false);
     expect(development.output?.inlineScripts).toBe(true);
-  });
-});
-
-const context: McpAppDiagnosticContext = Object.freeze({
-  appName: 'status',
-  entrySource: '/project/views/status.ts',
-  projectRoot: '/project',
-});
-
-const statsError = (overrides: Partial<Rspack.StatsError> & { readonly message: string }): Rspack.StatsError => ({ ...overrides });
-
-const swcSyntaxError = statsError({
-  code: 'ModuleBuildError',
-  message: '  × Module build failed (from builtin:swc-loader):\n  ╰─▶   × Syntax Error: Expression expected\n         ╭────\n       1 │ const x = ;\n         ·           ─\n         ╰────\n      \n',
-  moduleIdentifier: 'builtin:swc-loader??ruleSet[1].rules[2].oneOf[3].use[0]!/project/views/status.ts',
-  moduleName: './views/status.ts',
-});
-
-const unresolvedImportError = statsError({
-  loc: '1:1-41',
-  message: "  × Module not found: Can't resolve './missing-module' in '/project/views'\n   ╭─[1:0]\n 1 │ import { nope } from './missing-module';\n   · ────────────────────────────────────────\n 2 │ console.log(nope);\n   ╰────\n",
-  moduleIdentifier: 'builtin:swc-loader??ruleSet[1].rules[2].oneOf[3].use[0]!/project/views/status.ts',
-  moduleName: './views/status.ts',
-});
-
-describe('MCP App stats mapping', () => {
-  it('flattens an Rspack message to one line of prose', () => {
-    expect(normalizeStatsMessage(swcSyntaxError.message))
-      .toBe('Module build failed (from builtin:swc-loader): Syntax Error: Expression expected');
-    expect(normalizeStatsMessage(unresolvedImportError.message))
-      .toBe("Module not found: Can't resolve './missing-module' in '/project/views'");
-    expect(normalizeStatsMessage('  × Tsconfig not found /project/does-not-exist.json\n'))
-      .toBe('Tsconfig not found /project/does-not-exist.json');
-    expect(normalizeStatsMessage('\u001b[31mfailed\u001b[39m   badly\r\n\n  \u001b[2mdetail\u001b[22m'))
-      .toBe('failed badly detail');
-  });
-
-  it("locates an error from Rspack's loc, else the miette header, else the caret under the code frame", () => {
-    expect(statsErrorLocation(unresolvedImportError)).toEqual({ column: 1, line: 1 });
-    expect(statsErrorLocation(statsError({
-      message: '  × Syntax Error: Expression expected\n   ╭─[2:10]\n 1 │ export const a = 1;\n 2 │ const x = ;\n   ·           ─\n   ╰────\n',
-    }))).toEqual({ column: 10, line: 2 });
-    expect(statsErrorLocation(statsError({ message: '  × Syntax Error\n   ╭─[views/status.ts:3:4]\n' }))).toEqual({ column: 4, line: 3 });
-    // miette omits the header when the span starts on the first line.
-    expect(statsErrorLocation(swcSyntaxError)).toEqual({ column: 10, line: 1 });
-    expect(statsErrorLocation(statsError({ message: '  × Tsconfig not found /project/does-not-exist.json\n' }))).toBeUndefined();
-  });
-
-  it('resolves the failing module like Rsbuild does: file, then module name, then the loader chain target', () => {
-    expect(statsErrorFile(statsError({ file: 'views/a.ts', message: '', moduleName: './views/b.ts' }), '/project')).toBe('/project/views/a.ts');
-    expect(statsErrorFile(swcSyntaxError, '/project')).toBe('/project/views/status.ts');
-    expect(statsErrorFile(statsError({
-      message: '',
-      moduleIdentifier: 'builtin:swc-loader??ruleSet[1].rules[2].oneOf[3].use[0]!/elsewhere/views/status.ts?raw',
-    }), '/project')).toBe('/elsewhere/views/status.ts');
-    expect(statsErrorFile(statsError({ message: '', moduleIdentifier: '/project/views/a.css!=!builtin:lightningcss-loader!/project/views/a.css' }), '/project'))
-      .toBe('/project/views/a.css');
-    expect(statsErrorFile(statsError({ file: '', message: '' }), '/project')).toBeUndefined();
-  });
-
-  it('renders every stats error as an AB4770 and caps the list per App', () => {
-    expect(mcpAppCompileErrorDiagnostics(context, [swcSyntaxError, unresolvedImportError])).toEqual([
-      {
-        code: 'AB4770',
-        message: 'MCP App "status" failed to compile: views/status.ts:1:10: Module build failed (from builtin:swc-loader): Syntax Error: Expression expected',
-        recovery: 'Fix the reported error in the named file and rebuild; run `agent-bundle build` for the full message.',
-        severity: 'error',
-        sourcePath: '/project/views/status.ts',
-      },
-      expect.objectContaining({
-        message: "MCP App \"status\" failed to compile: views/status.ts:1:1: Module not found: Can't resolve './missing-module' in '/project/views'",
-      }),
-    ]);
-    // A module outside the project root shows absolutely; no module at all falls back to the entry.
-    expect(mcpAppCompileErrorDiagnostics(context, [
-      statsError({ message: '  × Module build failed\n', moduleName: '../shared/lib.ts' }),
-      statsError({ message: '  × Tsconfig not found /project/does-not-exist.json\n' }),
-    ])).toEqual([
-      expect.objectContaining({ message: 'MCP App "status" failed to compile: /shared/lib.ts: Module build failed', sourcePath: '/shared/lib.ts' }),
-      expect.objectContaining({
-        message: 'MCP App "status" failed to compile: Tsconfig not found /project/does-not-exist.json',
-        sourcePath: '/project/views/status.ts',
-      }),
-    ]);
-
-    const many = Array.from({ length: MCP_APP_COMPILE_ERROR_CAP + 3 }, (_, index) =>
-      statsError({ message: `  × failure ${String(index)}\n`, moduleName: './views/status.ts' }));
-    const capped = mcpAppCompileErrorDiagnostics(context, many);
-    expect(capped).toHaveLength(MCP_APP_COMPILE_ERROR_CAP);
-    expect(capped[MCP_APP_COMPILE_ERROR_CAP - 2]!.message).toContain(`failure ${String(MCP_APP_COMPILE_ERROR_CAP - 2)}`);
-    expect(capped[MCP_APP_COMPILE_ERROR_CAP - 1]).toEqual(expect.objectContaining({
-      code: 'AB4770',
-      message: 'MCP App "status" failed to compile: … and 4 more errors (run the compile with logLevel error via tools.rsbuild for the full list)',
-      sourcePath: '/project/views/status.ts',
-    }));
-    expect(mcpAppCompileErrorDiagnostics(context, many.slice(0, MCP_APP_COMPILE_ERROR_CAP))).toHaveLength(MCP_APP_COMPILE_ERROR_CAP);
-  });
-
-  it('renders warnings as AB4771 minus the ignore list', () => {
-    const warning = statsError({
-      loc: '3:1-40',
-      message: '  ⚠ Critical dependency: the request of a dependency is an expression\n',
-      moduleName: './views/status.ts',
-    });
-    const noise = statsError({ message: '  ⚠ \u001b[33mnoise\u001b[39m: something nobody can act on\n', moduleName: './views/status.ts' });
-    // Ignore patterns see the normalised text, so an entry reads like the message it silences.
-    expect(mcpAppCompileWarningDiagnostics(context, [warning, noise], [/^noise: /u])).toEqual([{
-      code: 'AB4771',
-      message: 'MCP App "status" produced a warning while compiling: views/status.ts:3:1: Critical dependency: the request of a dependency is an expression',
-      recovery: 'Address the reported warning in the named file and rebuild; run `agent-bundle build` for the full message.',
-      severity: 'warning',
-      sourcePath: '/project/views/status.ts',
-    }]);
-    // The default ignore list is empty: every warning surfaces.
-    expect(mcpAppCompileWarningDiagnostics(context, [warning, noise])).toHaveLength(2);
-    expect(mcpAppCompileWarningDiagnostics(context, [])).toEqual([]);
-  });
-
-  it('formats sizes 1024-based with one decimal, the same helper the CLI prints with', () => {
-    expect(formatByteSize(512)).toBe('512 B');
-    expect(formatByteSize(437_000)).toBe('426.8 KiB');
-    expect(formatByteSize(1_048_576)).toBe('1 MiB');
-    expect(formatByteSize(1_363_149)).toBe('1.3 MiB');
-  });
-
-  const statsModule = (overrides: Partial<Rspack.StatsModule> & { readonly size: number }): Rspack.StatsModule => ({
-    built: true,
-    buildTimeExecuted: false,
-    cached: false,
-    codeGenerated: true,
-    moduleType: 'javascript/auto',
-    sizes: { javascript: overrides.size },
-    type: 'module',
-    ...overrides,
-  });
-
-  const modules: readonly Rspack.StatsModule[] = [
-    // A concatenated module: only its parts are ranked, never its summed size.
-    statsModule({
-      identifier: '/project/views/status.ts + 2 modules',
-      modules: [
-        statsModule({ name: './views/status.ts', nameForCondition: '/project/views/status.ts', size: 2_048 }),
-        statsModule({ name: './views/StatusPanel.tsx', nameForCondition: '/project/views/StatusPanel.tsx', size: 4_096 }),
-        statsModule({ name: './.agent-bundle-virtual/generated/meta.mjs', nameForCondition: '/project/.agent-bundle-virtual/generated/meta.mjs', size: 128 }),
-      ],
-      name: './views/status.ts + 2 modules',
-      size: 6_272,
-    }),
-    statsModule({
-      name: '../../workspace/node_modules/.pnpm/react-dom@19.2.8_react@19.2.8/node_modules/react-dom/cjs/react-dom-client.production.js',
-      nameForCondition: '/workspace/node_modules/.pnpm/react-dom@19.2.8_react@19.2.8/node_modules/react-dom/cjs/react-dom-client.production.js',
-      size: 536_016,
-    }),
-    statsModule({ name: '../../workspace/node_modules/.pnpm/react@19.2.8/node_modules/react/cjs/react.production.js', nameForCondition: '/workspace/node_modules/.pnpm/react@19.2.8/node_modules/react/cjs/react.production.js', size: 17_217 }),
-    statsModule({ name: '../shared/theme.css', nameForCondition: '/shared/theme.css', size: 9_000 }),
-    statsModule({ moduleType: 'runtime', name: 'webpack/runtime/define property getters', size: 300 }),
-    statsModule({ name: '../../workspace/node_modules/.pnpm/scheduler@0.27.0/node_modules/scheduler/cjs/scheduler.production.js', nameForCondition: '/workspace/node_modules/.pnpm/scheduler@0.27.0/node_modules/scheduler/cjs/scheduler.production.js', size: 10_181 }),
-  ];
-
-  it('ranks the largest leaf modules under project-relative, node_modules, or absolute names', () => {
-    expect(largestModules(modules, '/project')).toEqual([
-      { name: 'node_modules/react-dom/cjs/react-dom-client.production.js', size: 536_016 },
-      { name: 'node_modules/react/cjs/react.production.js', size: 17_217 },
-      { name: 'node_modules/scheduler/cjs/scheduler.production.js', size: 10_181 },
-      { name: '/shared/theme.css', size: 9_000 },
-      { name: 'views/StatusPanel.tsx', size: 4_096 },
-    ]);
-    expect(largestModules(modules, '/project', 7).slice(5)).toEqual([
-      { name: 'views/status.ts', size: 2_048 },
-      { name: 'webpack/runtime/define property getters', size: 300 },
-    ]);
-  });
-
-  it('advises on a production view from 1 MiB and on any view past the host bound', () => {
-    const size = { bytes: 1_363_149, gzipBytes: 319_895 };
-    expect(mcpAppSizeDiagnostic(context, { mode: 'production', modules, size })).toEqual({
-      code: 'AB4772',
-      message: 'MCP App "status" compiled to 1.3 MiB (312.4 KiB gzip), above the 1 MiB advisory bound; largest modules: '
-        + 'node_modules/react-dom/cjs/react-dom-client.production.js (523.5 KiB), node_modules/react/cjs/react.production.js (16.8 KiB), '
-        + 'node_modules/scheduler/cjs/scheduler.production.js (9.9 KiB), /shared/theme.css (8.8 KiB), views/StatusPanel.tsx (4 KiB)',
-      recovery: 'Trim the largest modules listed and rebuild; the Workbench and serve-app hosts refuse a view above 2 MiB.',
-      severity: 'warning',
-      sourcePath: '/project/views/status.ts',
-    });
-    // The advisory is a production concern: readable development output is larger by design.
-    expect(mcpAppSizeDiagnostic(context, { mode: 'development', modules, size: { bytes: 1_572_864, gzipBytes: 400_000 } })).toBeUndefined();
-    expect(mcpAppSizeDiagnostic(context, { mode: 'production', modules, size: { bytes: MCP_APP_HTML_ADVISORY_BYTES - 1, gzipBytes: 1 } })).toBeUndefined();
-    expect(mcpAppSizeDiagnostic(context, { mode: 'production', modules: [], size: { bytes: MCP_APP_HTML_ADVISORY_BYTES, gzipBytes: 1 } })).toEqual(
-      expect.objectContaining({ message: 'MCP App "status" compiled to 1 MiB (1 B gzip), above the 1 MiB advisory bound' }),
-    );
-
-    const hostRefusal = ', above the 2 MiB bound the Workbench and serve-app hosts accept — the view will not render there';
-    for (const mode of ['development', 'production'] as const) {
-      const diagnostic = mcpAppSizeDiagnostic(context, { mode, modules, size: { bytes: MAX_APP_HTML_BYTES + 1, gzipBytes: 500_000 } });
-      expect(diagnostic?.code).toBe('AB4772');
-      expect(diagnostic?.message).toContain(`compiled to 2 MiB (488.3 KiB gzip)${hostRefusal}; largest modules: node_modules/react-dom`);
-    }
-    // Exactly the host bound still renders there; production still gets the advisory.
-    expect(mcpAppSizeDiagnostic(context, { mode: 'development', modules, size: { bytes: MAX_APP_HTML_BYTES, gzipBytes: 1 } })).toBeUndefined();
-    expect(mcpAppSizeDiagnostic(context, { mode: 'production', modules, size: { bytes: MAX_APP_HTML_BYTES, gzipBytes: 1 } })?.message)
-      .toContain('above the 1 MiB advisory bound');
   });
 });
